+++ src/libutil/rust-ffi.cc
